- rename classes - rearrange packages - formating
